renames files; reserves only `index.php` and new `crawler.php`
Deprecate old `cralwer.*` and `crawler_dfs.html`; rename
`crawler_dfs.php` to `crawler.php`.
