--- crawler.html
@@ -1,264 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-	<meta charset="utf-8">
-	<title>Facebook Data Crawler</title>
-	<script src="//connect.facebook.net/en_US/sdk.js"></script>
-	<script src="js/fbsdk-config.js"></script>
-	<script>
-		window.fbAsyncInit = function() {
-			FB.init(FBConfig);
-		};
-	</script>
-	<script src="js/fbsdk-extend.js"></script>
-	<script src="//code.jquery.com/jquery-2.1.4.min.js"></script>
-	<script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.4.7/angular.min.js"></script>
-	<script>
-		angular.module("myApp", [])
-		.controller("main", function($scope) {
-			$scope.connections = {
-				user: [
-					'albums', 'likes', //'feed', 'admined_groups', 'groups',
-					'posts', 'photos', 'tagged'//, 'videos'
-				],
-				page: [
-					'albums', 'events', //'feed',
-					'posts', 'photos', 'videos'
-				],
-				group: [
-					"albums", "events", "feed",
-					"members", "docs"//, "videos", "files"
-				],
-				event: [
-					"posts", "photos", "comments", "feed"//, "videos"
-				]
-			};
-
-			$scope.search = function() {
-				var q = $scope.q.trim();
-				$scope.cnadidates = [];
-				if(!q.length) return;
-				if($.isNumeric(q)) {
-					$scope.showSelect = false;
-					$scope.showInfo(q);
-				}
-				else {
-					$scope.message = 'Searching page ...';
-					$scope.candidates = [];
-					FB.apiwt("search?type=" + $scope.nodeType + "&q=" + q, function(r) {
-						if(r.data) {
-							$scope.message = '';
-							$scope.candidates = r.data;
-							$scope.showSelect = true;
-						}
-						else $scope.message = 'No results found.';
-						$scope.$apply();
-					});
-				}
-			};
-			$scope.showInfo = function(node_id) {
-				$scope.message = 'Fetching data ...';
-				$scope.nodeInfo = undefined;
-				$scope.conData = null;
-				FB.api(node_id + '?metadata=1', function(r) {
-					if(r.error) {
-						$scope.message = 'No such node.';
-						$scope.$apply();
-					}
-					else if(r.metadata.type != $scope.nodeType) {
-						$scope.message = 'Not a ' + $scope.nodeType + ' node.';
-						$scope.$apply();
-					}
-					else {	/// found a page node
-						var excludedFields = [
-							"token_for_business",
-							"currency",
-							"device",
-							"payment_pricepoints",
-							"viewer_can_send_gift",
-							"context",
-							"favorite_athletes",
-							"favorite_teams",
-							"ad_campaign",
-							"promotion_eligible",
-							"owner_business",
-							"access_token",
-							"business",
-							"can_checkin",
-							"can_post",
-							"last_used_time",
-							"has_added_app",
-							"is_permanently_closed",
-							"is_published",
-							"voip_info"
-						];
-						var ru = [];
-						for(var i = 0; i < r.metadata.fields.length; ++i) {
-							var fn = r.metadata.fields[i].name;
-							if($.inArray(fn, excludedFields) == -1) ru.push(fn);
-						}
-						ru = r.id + '?fields=' + ru.join(',');
-						FB.apiwt(ru, function(r) {
-							$scope.message = '';
-							$scope.nodeInfo = r;
-							if(!$scope.candidates.length) {
-								$scope.candidates.push(r);
-								$scope.nodeId = r.id;
-								$scope.showSelect = true;
-							}
-							$scope.$apply();
-						});
-					}
-				});
-			}
-			$scope.requestEdge = function(edge) {
-				$scope.message = 'Fetching data ...';
-				$scope.conData = null;
-				FB.apiwt($scope.nodeInfo.id + '/' + edge, function(r) {
-					$scope.message = '';
-					$scope.conData = r.data;
-					$scope.$apply();
-				});
-			};
-
-			$scope.getAlbums = function(nodeId) {
-				$scope.message = "Fetching data ...";
-				FB.apiwt(nodeId + "/albums", function(r) {
-					$scope.albums = r.data;
-					if(!r.data.length) $scope.message = '';
-					$scope.getAlbumInfo(r.data[0].id);
-				});
-			};
-			$scope.getAlbumInfo = function(nodeId) {
-				$scope.albumId = nodeId;
-				$scope.message = "Fetching data ...";
-				$scope.conData = null;
-				FB.apiwt(nodeId + "/photos", function(r) {
-					$scope.message = "";
-					$scope.conData = r.data;
-					$scope.$apply();
-				});
-			}
-
-			$scope.getGroups = function() {
-				$scope.clear();
-				FB.apiwt("me/groups", function(r) {
-					$scope.message = "";
-					$scope.candidates = r.data;
-					$scope.showSelect = true;
-					$scope.$apply();
-				});
-			}
-
-			$scope.clear = function() {
-				$scope.message = '';
-				$scope.q = '';
-				$scope.showSelect = false;
-				$scope.nodeInfo = undefined;
-				$scope.candidates = [];
-				$scope.albums = [];
-				$scope.conData = null;
-			};
-		});
-	</script>
-	<style>
-		ul { list-style-type: none; }
-		#message {
-			position: fixed;
-			top: 0;
-			right: 0;
-			margin: 0.2em;
-			padding: 0.2em;
-			background-color: #fff;
-		}
-		.cand_name {
-			display: inline-block;
-			border: 1px solid #ccc;
-			margin: 0.25em;
-			padding: 0.25em;
-		}
-		.info {
-			float: right;
-			width: 50%;
-			height: 20em;
-			white-space: pre-wrap;
-			font-family:
-			monospace;
-			overflow: scroll;
-		}
-		.edgeSelector {
-			min-height: 20em;
-		}
-		iframe {
-			width: 98%;
-			height: 40em;
-			border: solid #888;
-			border-width: 2px 0 0 0;
-		}
-	</style>
-</head>
-<body ng-app="myApp" ng-controller="main">
-	<h1>Facebook Data Crawler</h1>
-	<p id="message">{{message}}</p>
-	<h2>Choose what to crawl</h2>
-	<form method="get" action="crawler.php" target="iframe">
-		<dl>
-			<dt>Type</dt>
-			<dd>
-				<label><input type="radio" name="nodeType" ng-model="nodeType" value="user" ng-click="clear(); showInfo('me')">User</label>
-				<label><input type="radio" name="nodeType" ng-model="nodeType" value="page" ng-click="clear()">Page</label>
-				<label><input type="radio" name="nodeType" ng-model="nodeType" value="group" ng-click="getGroups()">Group</label>
-				<label><input type="radio" name="nodeType" ng-model="nodeType" value="event" ng-click="clear()">Event</label>
-			</dd>
-			<dt ng-show="nodeType=='page'||nodeType=='event'">{{nodeType}}</dt>
-			<dd ng-show="nodeType=='page'||nodeType=='event'">
-				<input ng-model="q" ng-change="search()" ng-model-options="{debounce: 500}"
-					placeholder="page ID or search text"
-				>
-			</dd>
-			<dt ng-show="showSelect">Select</dt>
-			<dd ng-show="showSelect">
-				<label class="cand_name" ng-repeat="cand in candidates">
-					<input type="radio" name="nodeId" value="{{cand.id}}"
-						ng-click="showInfo(cand.id);" ng-model="nodeId"
-					>{{cand.name}}
-				</label>
-			</dd>
-		</dl>
-		<div ng-show="conData||nodeInfo" class="info"
-		>{{(conData?conData:nodeInfo)|json}}
-		</div>
-		<h2>{{nodeInfo.name}}</h2>
-		<div class="edgeSelector">
-			<ul ng-show="nodeInfo">
-				<li ng-repeat="con in connections[nodeType]">
-					<label>
-						<input type="radio" name="edge" ng-value="con"
-							ng-click="$parent.albums=[]; requestEdge(con)"
-						>{{con}}
-					</label>
-				</li>
-				<li ng-hide="nodeType=='event'">
-					<label>
-						<input type="radio" name="edge" ng-value="'photosInAlbum'"
-							ng-click="getAlbums(nodeInfo.id)"
-						>photos in album
-					</label>
-					<ul>
-						<li ng-repeat="album in albums">
-							<label>
-								<input type="radio" name="album" value="{{album.id}}"
-									ng-click="getAlbumInfo(album.id)" ng-model="albumId"
-								>{{album.name}}
-							</label>
-						</li>
-					</ul>
-				</li>
-			</ul>
-			<button ng-show="nodeInfo" onclick="this.form.submit();"
-			>Start crawling</button>
-		</div>
-	</form>
-	<iframe name="iframe" src="about:blank"></iframe>
-</body>
-</html>
