--- crawler_dfs.html
@@ -1,58 +1,0 @@
-<!DOCTYPE HTML>
-<html ng-app="myApp">
-<head>
-	<meta charset="utf-8">
-	<title>Template</title>
-	<script src="//code.jquery.com/jquery-2.1.4.min.js"></script>
-	<script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.4.7/angular.min.js"></script>
-	<script>
-		angular.module("myApp", []).controller("main", function($scope, $http) {
-			$scope.timerId = 0;
-			$scope.interval = 3;
-			$scope.lastExecute = "Unknown";
-			$scope.path = "crawler_dfs.php";
-
-			$scope.request = function() {
-				$scope.lastExecute = (new Date).toISOString();
-				$http.get($scope.path).then(function(r){
-					$scope.data = r.data;
-					if($scope.timerId) $scope.timerId = window.setTimeout($scope.request, $scope.interval * 1000);
-				}, function(r){
-					$scope.data = JSON.stringify(r, undefined, 4);
-					if($scope.timerId) $scope.timerId = window.setTimeout($scope.request, $scope.interval * 5000);
-				});
-			}
-			$scope.keepRequest = function() {
-				$scope.request();
-				$scope.timerId = -1;
-			};
-			$scope.stop = function() {
-				if($scope.timerId > 0) window.clearTimeout($scope.timerId);
-				$scope.timerId = 0;
-			};
-			$scope.push = function() {
-				$scope.stop();
-				var tmp = $scope.path;
-				$scope.path = "crawler_dfs.php?type=page&path=/299951923544433";
-				$scope.request();
-				$scope.path = tmp;
-			};
-		});
-	</script>
-	<style></style>
-</head>
-<body ng-controller="main">
-	<h1 style="margin: 0;">Facebook Data Crawler</h1>
-	<button ng-click="push()">Push</button>
-	<br>
-	<button ng-click="request()" ng-disabled="timerId">Request once</button>
-	<br>
-	<button ng-click="keepRequest()" ng-disabled="timerId">Keep request</button>
-	in <input type="number" ng-model="interval" min="1" style="width: 4em;"> seconds.
-	<br>
-	<button ng-click="stop()" ng-disabled="!timerId">Stop</button>
-	<br>
-	Last execute: <span ng-bind="lastExecute"></span>
-	<div ng-bind="data" style="white-space: pre-wrap; font-family: monospace; border-top: 2px dashed #ccc;"></div>
-</body>
-</html>
